feat: multipart support for s3 backend

+++ src/Frontend/Objects/Delete.ts
@@ -12,6 +12,15 @@ export const deleteObject = (
     Effect.gen(function* () {
       const request = yield* HttpServerRequest.HttpServerRequest;
       const key = extractKey(request.url, bucket);
+      const url = new URL(request.url, "http://localhost");
+      const searchParams = url.searchParams;
+
+      if (searchParams.has("uploadId")) {
+        // Abort Multipart Upload
+        const uploadId = searchParams.get("uploadId")!;
+        yield* backend.abortMultipartUpload(key, uploadId);
+        return HttpServerResponse.empty({ status: 204 });
+      }
 
       yield* backend.deleteObject(key);
       return HttpServerResponse.empty({ status: 204 });
